Fix symbol handling

+++ shared/libblight_client/libc.h
@@ -10,10 +10,10 @@ namespace Libc {
   extern ssize_t (*write)(int, const void*, size_t);
   extern ssize_t (*writev)(int, const iovec*, int);
   extern ssize_t (*writev64)(int, const iovec*, int);
-  extern ssize_t (*pwrite)(int, const void*, size_t, int);
-  extern ssize_t (*pwrite64)(int, const void*, size_t, int);
-  extern ssize_t (*pwritev)(int, const iovec*, int, int);
-  extern ssize_t (*pwritev64)(int, const iovec*, int, int);
+  extern ssize_t (*pwrite)(int, const void*, size_t, off_t);
+  extern ssize_t (*pwrite64)(int, const void*, size_t, off64_t);
+  extern ssize_t (*pwritev)(int, const iovec*, int, off_t);
+  extern ssize_t (*pwritev64)(int, const iovec*, int, off64_t);
   extern int (*open)(const char*, int, ...);
   extern int (*open64)(const char*, int, ...);
   extern int (*openat)(int, const char*, int, ...);
